requirements.txt
fastapi
uvicorn
python-dotenv
openai
chromadb
langchain
langchain-community
tiktoken
langchain-openai